requirements.txt
easyocr==1.7.0
pandas==2.0.3
numpy==1.24.3
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
Pillow==9.5.0
fastapi
easyocr
pandas
fuzzywuzzy
uvicorn
python-multipart
transformers